+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [3.0.0-alpha3-SNAPSHOT] - 2020-09-21
-- added support for table visualization
+- added support for visualizing datasets as tables
 - extending markdown rendering to more types
 - bugfix in reporting the static rendering path